test(guest-download): reorganize integration coverage
