+++ crates/guest-download/tests/integration/main.rs
@@ -1,0 +1,6 @@
+mod support;
+
+mod binary_logging;
+mod download;
+mod file_scheme;
+mod scheduling;
